TT-725: make startup script work with versioned distributions
The startup script used to use distZip to build the application
distribution and then extract it and run it. As the archive would
contain a changing version number, it would be gnarly to accomodate
the version number in this script.

All of the above can be skipped by using the installDist task
instead.

Author: @tunylund