TT-854: Read metadata from hub and MSA
Note that MSA metadata is not signed, so it needs to be handled slightly differently.

We've implemented our own configuration classes that default almost everything to make it easier for users to configure.

Truststores are loaded from the resources folder unless paths are provided in the configuration.

There are separate healthchecks for hub and MSA metadata.

Authors: @richardtowers @georgievh @rachaelbooth