TT-858: Add scripts, test keys and certs, and stub msa metadata
The stub msa metadata is intended to be hosted on PaaS so we can run the
service provider without having to run an MSA as well.

The test keys and certs are intended to be used in tests only. The certs
have a lifespan of ~100 years, and they're hardcoded in the acceptance
tests and scripts. They're added here alongside the private keys for
reference.

Note that we're aware that these private keys are open source now - they
will not be used for anything sensitive.

Also added some scripts to generate keys and initialise the compliance
tool.

Authors: @richardtowers @georgievh