TT-1227: Provided entityIds are validated against a list in config.
The VSP will refuse to use entity Ids it has not been configured for.
It will return an error for either endpoint if it is provided with an
entity id which is not in config, or if it is not provided with an
entity id when there is more than one in config.

Authors: @rachaelbooth @sgreensmith