TT-1600 Remove defaults from config
Configuration file should not contain default values for
serviceEntityIds, environment, msaMetadata, or pki.
This removes them, and modifies startup.sh so it will look
for a local.env file and source it if present to provide
values for these.
The local.env file is git ignored, and should be generated
from verify-local-startup, or edited manually.

Authors: @rachaelbooth