bau: use a random port in tests
Tests can occasionally fail if something already has the VSP's default
port bound. We can mitigate this failure by overriding this value with 0
 in tests, which will give us a random port, .

There are some tests that require a specific port for testing the use of
environment variables, but they require a more complicated fix. Expect
a more complicated PR to follow.