rMMKhjYc: Run VSP in PaaS connected to integration
This adds a new deployment script and manifest specific to
integration connection. It removes the private keys from manifest
and set them as env variables. The keys are supplied as secrets by Jenkins.
The script along with the deployment pipeline should be refactored to be
simpler with less duplication - I raised a ticket for it
https://trello.com/c/MOIRqsXG/68-refactor-vsp-deployment-pipeline