Remove information about support for compilation on Java 9 and 10
This feature was built in May 2018, with the intention of releasing the 
VSP soon after, which didn't happen.

SMEs say it's better to not bring Java 9 and 10 into the discussion 
because they are end-of-life.

In any case, SMEs said users are not likely to compile their own VSP. 
They'll use what we provide.