YmnUf9Ma: create a distribution for testing on the PaaS
When we test on the PaaS we need to run the VSP is variety of
different modes:
- VSP connected to a MSA
- VSP standalone with eIDAS enabled
- VSP standalone with eIDAS disabled

As MSA support is defined using the yml file we can provide the VSP
that we supply to PaaS with a set of different yml files.
Using gradle's distribution plugin it's possible to create
a testing distribution that contains the files that are needed.