fix: support for eidas identities in development mode
When the VSP runs in development mode we need to be able to have some
support for eIDAS identities. An eIDAS identity will have attributes that
are missing `from` and `to` dates. Some attributes may also be absent.

This change includes tests and changes to the how the Dto that is sent
to CT, and contains the identity, serializes itself.