EID-1325: Add RequestedAuthnContext to AuthnRequest
The `levelOfAssurance` parameter sent to the `/generate-request` endpoint
was being ignored. This change adds the corresponding AuthnContextClassRef(s) to
the RequestedAuthnContext of the AuthnRequest generated by the VSP.

In the case of LOA1, we need to add both the level 1 and level 2
AuthnContextClassRefs while for LOA2 we only need the level 2
AuthnContextClassRef.