EID-1605 Validate signatures of country signed Responses
To be able to handle unsigned assertions from certain countries we need
the VSP to accept the original response from the country, rather than a
response from the hub.

The logic for validating country signatures has been transplanted from
the hub into saml-lib, and then used here in the ResponseService.

In the case that the VSP is being used with an MSA (matching), the VSP
doesn't handle eIDAS attributes.
In the case that the VSP is on it's own, it can either be eIDAS enabled
or not.
This commit allows a ResponseService to be configured either way, and to
validate signatures appropriately.