Initial commit implementing an API for interacting with Verify via OIDC.
There are currently two endpoints a service could use - one to get an
authorisation code and one to swap the authorisation code for claims from
the token and userinfo endpoints. Some more thought is needed around
what this interface should look like for a service user. It's currently
hardcoded to talk to prototype endpoints on verify-stub-idp.