EID-1607 Add EidasUnsignedAssertionTranslator
In an eIDAS enabled non-matching set up, the `assertionTranslator` used
by the `ResponseService` is a `ClassifyingAssertionTranslator`.

This translator is configured with different translators and will
determine which to use depending on the assertions given to it.

This extends it to include the new `EidasUnsignedAssertionTranslator`
which will deal with unsigned assertions.

The only difference in dealing with signed and unsigned assertions is
the validation of the signature. As such, the vast majority of the logic
in `EidasAssertionTranslator` has been moved to an abstract base class
which is extended by the new `EidasAssertionTranslator` as well as the
new one. They both just implement the abstract `validateSignature`
method.

The implementation for the unsigned translator is simply a no-op.