EID-1607 Add an acceptance test for unsigned assertions
This test creates an Eidas Response with an unsigned encrypted
assertion. It then converts that Response into a string and stuffs it
into a hub Response.

The hub Response has two attributes in its assertion. One is the
original eIDAS response, the other is the encryption keys used in that
response, which have been extracted and encrypted for the RP.

The encryption algorithm used in the eIDAS response has been changed to
be one of the eIDAS spec algorithms.