EID-1762 Move dto's and logic to saml-lib
We want to consolidate all the logic for logging hashes of eIDAS
attribtues. We log in the proxy-node and the hub.

The proxy-node receives attributes over the wire from the VSP. This
moves relevant dto's associated with this to saml-lib so proxy-node can
de-serialize into them. This ensures we have the same objects.

We also move the `MatchingDatasetToNonMatchingAttributesMapper` to
saml-lib. This is so that the hub can use it to extract attributes from
assertions for logging in the same way the VSP does.