BAU: Make it easier to run the VSP locally
The VSP can be run locally without providing the `environment` config value and instead overriding the Hub endpoints and truststores.
However, when those values are provided but `environment` is null, VSP would fail to start up due to null pointer exceptions. This is because in statements with Optionals like this `ofNullable(x).orElse(y)` Java evaluates the `orElse` bit even if x is not null - it doesn't short circuit the OR logic. However, if the same is changed to this form - `ofNullable(x).orElseGet(() -> y)` - the lambda function is only evaluated if x is indeed null.

Author: @CharlesIC