BAU: Revert including local binaries in docker workspace
This used to be so we could build local docker images of the VSP using binaries built locally instead of building them in docker every time.

This used to be needed for the Proxy Node local startup to work but now we mount the bins so they don't need to be included in Docker context. This makes docker build run faster.

Author: @CharlesIC